init the Musik and Light control board repo
Musik and Light control board belongs to the project “Waagschalspiel”
for Leuchtenburg.
It control for LED, a music board that have the ability to play .wav
files which can have up to 8 channel.
It use “Perlin noise” to simulate candle effect. It can blend the sound
volume for each channel and calculate the accrual speed on each Nanotec
PD4-N step motors. In order to get the right sound effect for different
Motor speed.